Added the setup for the output functionalities including the generate_heatmap function
