+++ landslideml/__init__.py
@@ -3,3 +3,4 @@ Basic package for landslide susceptibility modeling using machine learning.
 """
 from .reader import generate_model, load_model
 from .model import MlModel
+from .output import generate_heatmap
